Fix 0-length collision responses resulting in black screen
The issue happened occasionally when the ball was against the board
bottom or right wall and the game was in progress, and always when the
ball was in those positions and the game was paused and then resumed.

Normalizing the 0-length response vector when applying the collision
response resulted in NaNs, which propagated through ball velocity to
ball position vector. This resulted in invalid view matrix yielding the
black screen.